pyproject.toml
[project]
name = "jw-org-mcp"
version = "0.1.0"
description = "Model Context Protocol server for verified access to jw.org content"
readme = "README.md"
authors = [
    { name = "Bjern Potgieter", email = "bjern@toxzen.co.za" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "brotli>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[project.scripts]
jw-org-mcp = "jw_org_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=jw_org_mcp --cov-report=term-missing --cov-report=html"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]
